# Optimize cast(uuid as varchar)