Summary:
boost::lexical_cast used to implement cast(uuid as varchar) is very slow.

Replace with custom optimization.

Microbenchmark shows 20x improvement.

The benchmark compares uuid() with cast(uuid() as varchar). The latter includes the code of the former plus the cost of the cast. Before the change, uuid() + cast was 16s, 10s more than uuid() alone. After the change, uuid() + cast() is just 0.5s more.

Before:

```
============================================================================
[...]hmarks/ExpressionBenchmarkBuilder.cpp     relative  time/iter   iters/s
============================================================================
cast##no_cast                                                6.33s   157.86m
cast##as_varchar                                            16.54s    60.45m
```

After:

```
============================================================================
cast##no_cast                                                6.29s   159.10m
cast##as_varchar                                             6.81s   146.74m
----------------------------------------------------------------------------
```

Profile before the optimization:  {F1735048022}

Differential Revision: D59229653